Move InstanceUtilities test to separate file.
